fix(02): correct maxDiff
I misread the problem as a diff of 3 being invalid, when that is the
max diff allowed. Luckily, I made that a nice const and the change is
super simple.

I decided to update the tests to be slightly more robust, as they were
right at the limit. This isn't perfect, as a maxDiff of 99 will break
again, but it is good enough. I could go find how to represent the min
and max int values, which should be complete.