feat(02): try removing before and after diff
I added tests for the dampened function. It turns out that there was
more wrong than I thought. The second test case was also failing. For
the sequence [3, 0, 2, 1] the reported failure was at index 2. Removing
that yields [3, 0, 1] which is also invalid. Attempting to remove both
before and after means I also try [3, 2, 1] which is valid.

However, there is still a problem. [0, 3, 2, 1] does not report a
problem until index 2 as well, so I only end up trying [0, 2, 1] and
[0, 3, 1], both of which are invalid. I am going to try adding one more
check, for the previous value, but I think this means that I should
generally change strategy.