fail(02): attempt to check previous
I tried to remove the previous value. However, it turns out that these
slice and append operations do not always give you a new slice. I can
see why that would be useful, but here it leads to weird bugs. This now
fails the second and third test cases because the slice is modified.
This is presumably impacting the other pair of calls as well.

This are the kinds of failures I want to see. I am using this to learn
go, and this type of thing is what I need to know about.