feat(03): add sumMulsConditional
This took a bit longer than expected. I kept tripping myself up by
adding values to the result of strings.Index, which means that -1 isn't
the correct sentinel value.

A small test suite and some research into dlv made it pretty easy to
figure out where the issuers were.

I think my very first idea, which was to add the on and off commands to
the regex and just update a bool would have been much faster to
implement.

It is my understanding that these slices are cheap to construct, and the
solution I created here isn't actually copying memory, but is updating
indices.