feat(06): add ability to parse input into a field and starting position
The choice to not mark the starting position is arbitrary, but does need
to be known. Later when the simulation runs I will have to mark where
the robot was, not where it ends up. The biggest difference I see is the
last step, before the robot walks off the map. This might make that part
easier, but it may also make other things harder, so for now I don't
have a preference.

I also note that creating a Field feels kind of odd. Having to specify
one length in make isn't very clear, it probably would be nicer with a
separate function to instantiate it.