refactor(06): restructure simulation loop
Each iteration of the loop used to represent the action of leaving a
cell. In addition, turning was represented by updating the direction but
not the position of the robot.

Looking for loops requires us to look one step ahead, which is difficult
when loops don't always mean the same thing. You could make it work, but
the work gathers in step forward case.