test(06): add separate loops test
I thought this was going to expose a hole in the algorithm, but I think
I am ahead of myself. I thought that after the 6th turn, it would try to
place a wall at 5,3, which would block the original path. However, it
doesn't do that because the visited direction is to the right not the
left.

This will something that could go wrong once we trying simulating
placing blocks ahead of the robot, so the test is still valuable.