refactor(06): extract getNextDir
This function is a shocking amount of the work done in the simulation
loop. I am working towards having a "virtual" simulation that just finds
the old path but isn't allowed to place walls. This will support finding
loops that are not caused by placing a wall at the intersections with
the old path.