fix(09): correct end recovery walk direction
When decrementing the end position, we need to skip empty blocks. I had
accidentally incremented end instead of decrementing it. This has the
effect of only ever using the last block's id for any empty block.

This was even exposed in my original test case, but I didn't validate by
hand. I just assumed I was correct in my implementation and took the
value it gave back.

Solving this bug gets the correct answer for the puzzle.