fail(09): compute from disk map directly
This was an interesting exercise, but ultimately does not get the right
result. I was able to play around with parametereized tests that test
each function with all the given inputs, which is neat. The second
version passes all the tests I have, but does not give the same result
for the puzzle input.